Dashboard, DrawingSession and Explore updated
Dashboard updated: switched explore and painting sections, added a drawing session and deleted the search bar.

DrawingSession added: now you can draw hand-free.

Light and Dark mode added: must be adjsuted the design but it's fully functional and must be decided if we need the theme picker in the explore section or in the dashboard(preferred because it is always accessible).

MyPaintingView: created the MyPaintingView where the saved paintings will be shown (not functional atm).